Refactor application structure by introducing AppClassicKomet and AppMenu classes, and updating AppInterface methods for better modularity and maintainability.
