+++ application/src/main/java/dev/ikm/komet/app/AppMenu.java
@@ -1,0 +1,72 @@
+package dev.ikm.komet.app;
+
+import javafx.application.Platform;
+import javafx.scene.control.Menu;
+import javafx.scene.control.MenuBar;
+import javafx.scene.control.MenuItem;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyCodeCombination;
+import javafx.scene.input.KeyCombination;
+import javafx.scene.layout.BorderPane;
+import javafx.stage.Stage;
+
+import static dev.ikm.komet.app.WebApp.IS_BROWSER;
+
+public class AppMenu {
+
+    private final AppInterface app;
+
+    public AppMenu(AppInterface app) {
+        this.app = app;
+    }
+
+    void generateMsWindowsMenu(BorderPane kometRoot, Stage stage) {
+        MenuBar menuBar = new MenuBar();
+        Menu fileMenu = new Menu("File");
+
+        MenuItem about = new MenuItem("About");
+        about.setOnAction(_ -> app.showAboutDialog());
+        fileMenu.getItems().add(about);
+
+        // Importing data
+        MenuItem importMenuItem = new MenuItem("Import Dataset");
+        importMenuItem.setOnAction(actionEvent -> app.openImport(stage));
+        fileMenu.getItems().add(importMenuItem);
+
+        // Exporting data
+        MenuItem exportDatasetMenuItem = new MenuItem("Export Dataset");
+        exportDatasetMenuItem.setOnAction(actionEvent -> app.openExport(stage));
+        fileMenu.getItems().add(exportDatasetMenuItem);
+
+        MenuItem menuItemQuit = new MenuItem("Quit");
+        KeyCombination quitKeyCombo = new KeyCodeCombination(KeyCode.Q, KeyCombination.CONTROL_DOWN);
+        menuItemQuit.setOnAction(actionEvent -> app.quit());
+        menuItemQuit.setAccelerator(quitKeyCombo);
+        fileMenu.getItems().add(menuItemQuit);
+
+        Menu editMenu = new Menu("Edit");
+        MenuItem landingPage = new MenuItem("Landing Page");
+        KeyCombination landingPageKeyCombo = new KeyCodeCombination(KeyCode.L, KeyCombination.CONTROL_DOWN);
+        landingPage.setOnAction(actionEvent -> app.launchLandingPage(app.getPrimaryStage(), null /* userProperty.get()*/));
+        landingPage.setAccelerator(landingPageKeyCombo);
+        landingPage.setDisable(IS_BROWSER);
+        editMenu.getItems().add(landingPage);
+
+        Menu windowMenu = new Menu("Window");
+        MenuItem minimizeWindow = new MenuItem("Minimize");
+        KeyCombination minimizeKeyCombo = new KeyCodeCombination(KeyCode.M, KeyCombination.CONTROL_DOWN);
+        minimizeWindow.setOnAction(event -> {
+            Stage obj = (Stage) kometRoot.getScene().getWindow();
+            obj.setIconified(true);
+        });
+        minimizeWindow.setAccelerator(minimizeKeyCombo);
+        minimizeWindow.setDisable(IS_BROWSER);
+        windowMenu.getItems().add(minimizeWindow);
+
+        menuBar.getMenus().add(fileMenu);
+        menuBar.getMenus().add(editMenu);
+        menuBar.getMenus().add(windowMenu);
+        //hBox.getChildren().add(menuBar);
+        Platform.runLater(() -> kometRoot.setTop(menuBar));
+    }
+}
